un peu mieux ca

+++ srcs/Windows/move.c
@@ -6,7 +6,7 @@
 /*   By: bmangin <bmangin@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/18 11:35:27 by bmangin           #+#    #+#             */
-/*   Updated: 2021/05/18 14:59:16 by bmangin          ###   ########lyon.fr   */
+/*   Updated: 2021/05/21 17:26:46 by bmangin          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,19 +14,19 @@
 
 void	mov_updown(t_g *g, int neg)
 {
-	int mov;
+	int		mov;
 
 	mov = SPEED * neg;
-	if (!is_wall(*g, g->player.x, g->player.y + mov))
+	if (!is_wall(*g, (int)(g->player.x), (int)(g->player.y + mov)))
 		g->player.y += (SPEED * neg);
 }
 
 void	mov_lateral(t_g *g, int neg)
 {
-	int mov;
+	int		mov;
 
 	mov = SPEED * neg;
-	if (!is_wall(*g, g->player.x + mov, g->player.y))
+	if (!is_wall(*g, (int)(g->player.x + mov), (int)(g->player.y)))
 		g->player.x += (SPEED * neg);
 }
